Drop the loopback module, handle loop streams directly.
The loopback module tries to dynamically re-sample and buffer the audio
to maintain a constant latency. This is *horrible* behavior when the
input is not perfectly reliable such as A2DP streams. The result is
frequent audible pitch bends and temp changes as the sample rate is
pushed to high or low in order to fill or empty the buffer. gack.

For A2DP high latency is ok, changing sample rates are unacceptable, and
if the buffer runs out I will live with brief audio gaps. This is pretty
much how normal A2DP devices work.

This first stab is very minimal and does no buffering beyond what pulse
does internally. Eventually some form of buffering will be required.