Connect to BlueZ over DBus to get stream state info.
The original intent here was to flush the audio streams as soon as
playback was stopped. That way the user would get immediate stops rather
than waiting for the now much larger buffer to run out. However the
AudioSource interface is too slow in reporting the state change, perhaps
because it is simply waiting for the stream to go idle itself. Dunno.